Cambios en transbank

+++ app/Http/Controllers/CartController.php
@@ -255,19 +255,23 @@ class CartController extends Controller
 
 
     // ORDEN CONFIRMADA FINALMENTE
-    public function confirmOrder($orderId)
+    public function confirmOrder(Request $request, $orderId)
     {
-        $user = Auth::user();
-        $order = Order::findOrFail($orderId);
-       
-        if ($user && $order->user_id === $user->id && $order->estado === 0) {
-            $order->estado = 1; // Cambiar el estado a pagado
-            $order->save();
-            Mail::to($user->email)->send(new ProofPayment($order->id));
-            Cart::destroy();
-            return redirect()->route('home-landing')->with('success', 'La Compra se realizó correctamente');
+        if ($request->isMethod('post')) {
+            $user = Auth::user();
+            $order = Order::findOrFail($orderId);
+           
+            if ($user && $order->user_id === $user->id && $order->estado === 0) {
+                $order->estado = 1; // Cambiar el estado a pagado
+                $order->save();
+                Mail::to($user->email)->send(new ProofPayment($order->id));
+                Cart::destroy();
+                return redirect()->route('home-landing')->with('success', 'La Compra se realizó correctamente');
+            } else {
+                return redirect()->back()->with('error', 'No se puede confirmar la orden');
+            }
         } else {
-            return redirect()->back()->with('error', 'No se puede confirmar la orden');
+            return abort(405); // Método no permitido (GET)
         }
     }
 }
